Bank App With Flutter
This application was implemented with flutter. There is a basic login page, the user registers and this information is stored in the database. The user's account information is available on the home page. You can buy and sell foreign currency and perform the necessary transactions in the relevant menus. It does not work like a real bank application, it is just an attempt to emulate it.